Fixed issues with code example in README
These changes were made to the wiki home page during the release process
for v1.2.0 but the README was not updated with the same changes.
I have updated the sub-tasks of the cloneable release task to include
a mention to update the README code samples and also moved the sub-task
named 'Update the wiki home page and README' from step 11 to step 7, so
the README will be updated before merging develop into master.